Create a new 'hadoop' module, and move non-Spark code into this from the 'spark' module
