+++ examples/src/test/scala/dev/chronicles/examples/DatePartitionedTableLoaderSpec.scala
@@ -8,7 +8,7 @@ import dev.chronicles.core.Partition.PartitionColumn
 import dev.chronicles.core.VersionTracker._
 import dev.chronicles.core._
 import dev.chronicles.spark.SparkHiveSuite
-import dev.chronicles.spark.filesystem.VersionedFileSystem
+import dev.chronicles.hadoop.filesystem.VersionedFileSystem
 import org.scalatest.{FlatSpec, Matchers}
 
 /**
