+++ hadoop/src/test/resources/log4j.properties
